% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_data.R
\name{api_data}
\alias{api_data}
\alias{api_data_1}
\alias{api_data_2}
\alias{api_data_12}
\title{PX-Web Data by API}
\usage{
api_data(
  url_or_tableid,
  ...,
  url_type = "ssb",
  use_index = FALSE,
  default_query = c(1, -2, -1),
  return_dataset = NULL,
  make_na_status = TRUE,
  response_format = "json-stat2",
  verbose_print = FALSE
)

api_data_1(..., return_dataset = 1)

api_data_2(..., return_dataset = 2)

api_data_12(..., return_dataset = 12)
}
\arguments{
\item{url_or_tableid}{A table id, a PxWebApi 2 URL to data or metadata, or metadata returned by
\code{\link[=meta_data]{meta_data()}} or \code{\link[=meta_frames]{meta_frames()}}.}

\item{...}{Specification of query for each variable.}

\item{url_type}{Currently two possibilities: "ssb" (Norwegian) or "ssb_en" (English).}

\item{use_index}{Logical. If TRUE, numeric values are matched against the \code{index} variable in
the metadata, which usually starts at 0. If FALSE (default), numeric values
are interpreted as row numbers in the metadata, using standard R indexing.
Negative values can be used to specify reversed row numbers.}

\item{default_query}{Specification for variables not included in \code{...}}

\item{return_dataset}{Possible non-NULL values are \code{1}, \code{2} and \code{12}. Then a single data set is returned as a data frame.}

\item{make_na_status}{When TRUE and when dataPackage is \code{"rjstat"} and when missing entries in \code{value},
the function tries to add an additional variable, named \code{NAstatus}, with status codes.
An explanation of these status codes is provided in the note part of the comment attribute,
i.e. what you get with \code{\link[=note]{note()}}. See the bottom example.}

\item{response_format}{Response format to be used when \code{apiPackage} and \code{dataPackage} are defaults  (\code{"json-stat"} or \code{"json-stat2"}).}

\item{verbose_print}{When TRUE, printing to console}
}
\value{
A data frame, or a list of data frames, depending on the input and
parameters.
}
\description{
This function constructs a PxWebApi 2 data URL using \code{\link[=query_url]{query_url()}}
and retrieves the data using \code{\link[=get_api_data]{get_api_data()}}.
}
\examples{

 api_data(14162, Region = FALSE, InnKvartering1 = FALSE, Landkoder2 = FALSE, 
          ContentsCode = TRUE, Tid = 3i, url_type = "ssb_en")

 # same as above 
 # api_data("https://data.ssb.no/api/pxwebapi/v2/tables/14162/data?lang=en", 
 #           default_query = FALSE, ContentsCode = "*", Tid = "top(3)")

 api_data_1("https://data.ssb.no/api/pxwebapi/v2/tables/09546/data?lang=en",
            Region = FALSE, SkoleSTR = "07", GrSkolOrgForm = "4", 
            EierforhSkole = 1:2, ContentsCode = TRUE, Tid = 5i) 

}
