% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PxWebApi2.R
\name{make_url}
\alias{make_url}
\title{Make data url from table id}
\usage{
make_url(id, url_type = "ssb")
}
\arguments{
\item{id}{Table id (integer).}

\item{url_type}{Currently two possibilities: "ssb" (Norwegian) or "ssb_en" (English).}
}
\value{
A character string (URL).
}
\description{
Make data url from table id
}
\examples{
make_url(4861)
make_url(4861, "ssb_en")
}
\keyword{internal}
