% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApiData.R
\encoding{UTF8}
\name{MakeUrl}
\alias{MakeUrl}
\title{MakeUrl from id}
\usage{
MakeUrl(id, urlType = "SSB")
}
\arguments{
\item{id}{integer}

\item{urlType}{Currently two possibilities: "SSB" (Norwegian) or "SSBen" (English)}
}
\value{
url as string
}
\description{
MakeUrl from id
}
\examples{
MakeUrl(4861)
MakeUrl(4861, "SSBen")
}
\keyword{internal}
