% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PxWebApi2.R
\name{metadata_url}
\alias{metadata_url}
\title{Make metadata url from data url}
\usage{
metadata_url(url)
}
\arguments{
\item{url}{URL to data.}
}
\value{
A character string (URL).
}
\description{
Make metadata url from data url
}
\examples{
metadata_url(make_url(4861))
metadata_url(make_url(4861, "ssb_en"))
}
\keyword{internal}
