% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PxWebApi2.R
\name{metadata_url}
\alias{metadata_url}
\title{Convert between data and metadata URLs}
\usage{
metadata_url(url, data_url = FALSE)
}
\arguments{
\item{url}{A PxWebApi URL to data or metadata.}

\item{data_url}{Logical. When TRUE, return a data URL. When FALSE (default),
return a metadata URL.}
}
\value{
A character string (URL).
}
\description{
Converts a PxWebApi v2 data URL to a metadata URL, or vice versa.
}
\examples{
metadata_url(make_url(4861))
metadata_url(make_url(4861, "ssb_en"), data_url = TRUE)
}
\keyword{internal}
